This release introduces breaking changes, be careful when updating.
Users now have usernames, adding admins does not depend on positions
anymore. This is way easier to use and more convenient. Requires client
& server configuration update.

Configuration has been updated:
- USER_TOKENS has been removed in favor of USERS (new format)
- MAX_ADMIN_ID has been removed in favor of ADMINS (new format)
- DELETION_SALT has been renamed to SALT (same format)

Some parameters have been changed:
- `auth_token` has been renamed to `token`
- `deletion_hash` has been renamed to `key`
- `user_id` has been removed in favor of `username` and now only exists
when a user is authenticated

As of now, this is a beta version.
A migration guide will be added for the release.

Version bumped to 2.0.0-beta1