Patch security issue on empty salt
A configured salt is necessary for generating security keys
According to the comment above its definition, keeping it empty
was supposed to disable the feature. Instead, it kept it insecurely
enabled and used the empty salt, giving absolutely predictible
security keys considered as valid instead of not returning any
Note that commenting-out the define properly disabled this feature

This behavior is now fixed. If you do not update and have an empty
salt, either remove/comment the define, or set one immediately

Default allowed file extensions have also been edited, showing
regex are supported (which they are since the beginning) to make
sure users are aware of that possibility as it wasn't documented

Version bumped to 2.1.1