Add URL_STRIP_EXTENSION constant (#2)
This small update allows users to remove the file extension from
the resulting URL, as requested. This is useful for instances using
a rewrite rule in order to automatically find the corresponding file.

Be careful, as the filename generator only checks against existing
files of the same extension. It is therefore possible to have two files
with the exact same name, as long as they have a different extension.

Version bumped to 2.3.0